Refuse to show source of 'View Source' page
The size of the URI to open becomes very large even at just one level of
source viewing recursion. Having a lot of browsing history is a serious
performance issue in iOS, to an extent that we may even have to abandon
the current URI implementation for this extension to be practically
usable. This is a simple quick fix which is better than nothing.